Allow double indent with indent_func_*_param=false.
Make it possible to set indent_func_param_double=true even with
indent_func_*_param set to false, such that the double indent is applied
to code like

void very_long_function_name(
        int param,
        int param2)

The above case is now handled the same with both values of
indent_func_*_param, except for use of tabs/spaces. The difference is
only to preserve existing functionality; may not be the best solution in
the long run.

Test 30109 still fails, but it seems to be partly an unrelated issue.